[Feat] Pull to Refresh 새로 고침 구현하기 & 상단 뷰 분리
